Initial commit, after refactoring the old code.
The original code is still available at http://hg.avaraline.net/javara/. The code is quite a bit cleaner now, but I still think other improvements can be made. I relied quite a bit on the attachedToWorld callback, but it might be better to shift that code to the worldInitialized callback in most cases. Additionally, update methods should be written for each WorldObject type that keeps colors and so forth updated in case they were changed during gameplay. I think I'd like to rework the WorldObject subclass constuctors a bit, too.

A number of things don't work post-refactor. My goal for this phase was to get map loading ironed out and more consistent/predictable. I have yet to add sound, grenades, or goodies back in, as jME3 appears to have changed the way collision listeners worked from before. The Hector models also don't seem to load, and I removed crouching in order to get the code to run, but the Hector and Player classes basically need to be rewritten from scratch, anyway. Incarnators only ever face in the direction of North (toward negative Z), but this actually seemed to be the case in the code pre-refactor with the latest jME3 nightlies. I've also noted an issue where the SkyPlane quad (which draws the gradient SkyBox) occasionally "shrinks" but am unable to reproduce it reliably.